Refactor menu and add create simple factory.
