+++ src/main/java/com/study/pattern/factory/abstracts/po/BoyInterface.java
@@ -1,0 +1,14 @@
+package com.study.pattern.factory.abstracts.po;
+
+/**
+ * 男孩接口
+ *
+ * @author Sky
+ * @Date 2019-01-27 14:12.
+ */
+public interface BoyInterface {
+    /**
+     * 画一个男孩
+     */
+    void drawMan();
+}
