+++ src/main/java/com/study/pattern/factory/abstracts/po/GirlInterface.java
@@ -1,0 +1,14 @@
+package com.study.pattern.factory.abstracts.po;
+
+/**
+ * 女孩接口
+ *
+ * @author Sky
+ * @Date 2019-01-27 14:13.
+ */
+public interface GirlInterface {
+    /**
+     * 画一个女孩
+     */
+    void drawWomen();
+}
