+++ src/main/java/com/study/pattern/factory/methods/po/HairLeft.java
@@ -1,0 +1,14 @@
+package com.study.pattern.factory.methods.po;
+
+/**
+ * 左偏分发型
+ *
+ * @author Sky
+ * @Date 2019-01-27 13:03.
+ */
+public class HairLeft implements HairInterface {
+    @Override
+    public void draw() {
+        System.out.println("-----------------左偏分发型-------------------");
+    }
+}
