+++ src/main/java/com/study/pattern/factory/methods/type.properties
@@ -1,0 +1,3 @@
+left = com.study.pattern.factory.methods.po.HairLeft
+right = com.study.pattern.factory.methods.po.HairRight
+in = com.study.pattern.factory.methods.po.HairIn
