+++ src/main/java/com/study/pattern/factory/工厂模式常见应用.jpg
