+++ src/main/java/com/study/pattern/factory/类图-工厂方法模式类图.jpg
