+++ src/main/java/com/study/pattern/factory/类图-抽象工厂模式类图.jpg
