+++ src/main/java/com/study/pattern/factory/需求及应用场景.jpg
