+++ src/main/java/com/study/pattern/graphic/creational/factory/simple/Product.java
@@ -1,0 +1,12 @@
+package com.study.pattern.graphic.creational.factory.simple;
+
+/**
+ * 抽象产品角色
+ * 抽象产品角色是所创建的所有对象的父类，负责描述所有实例所共有的公共接口
+ *
+ * @author Sky
+ * @date 2019-11-01 16:23.
+ */
+public interface Product {
+    void user();
+}
