+++ src/main/java/com/study/pattern/graphic/creational/factory/simple/SimpleFactory.java
@@ -1,0 +1,12 @@
+package com.study.pattern.graphic.creational.factory.simple;
+
+/**
+ * 简单工厂模式
+ * 工厂角色负责实现创建所有实例的内部逻辑
+ *
+ * @author Sky
+ * @date 2019-11-01 16:22.
+ */
+public class SimpleFactory {
+
+}
