--- src/main/java/com/study/patterns/factory/abstracts/ChristmasFactory.java
@@ -1,24 +1,0 @@
-package com.study.patterns.factory.abstracts;
-
-import com.study.patterns.factory.abstracts.po.BoyInterface;
-import com.study.patterns.factory.abstracts.po.ChristmasBoy;
-import com.study.patterns.factory.abstracts.po.ChristmasGirl;
-import com.study.patterns.factory.abstracts.po.GirlInterface;
-
-/**
- * 圣诞系列加工厂
- *
- * @author Sky
- * @Date 2019-01-27 14:18.
- */
-public class ChristmasFactory implements PersonFactory {
-    @Override
-    public BoyInterface getBoy() {
-        return new ChristmasBoy();
-    }
-
-    @Override
-    public GirlInterface getGirl() {
-        return new ChristmasGirl();
-    }
-}
