--- src/main/java/com/study/patterns/factory/abstracts/PersonFactory.java
@@ -1,24 +1,0 @@
-package com.study.patterns.factory.abstracts;
-
-import com.study.patterns.factory.abstracts.po.BoyInterface;
-import com.study.patterns.factory.abstracts.po.GirlInterface;
-
-/**
- * 获取人物接口
- *
- * @author Sky
- * @Date 2019-01-27 14:11.
- */
-public interface PersonFactory {
-    /**
-     * 男孩接口
-     * @return
-     */
-    BoyInterface getBoy();
-
-    /**
-     * 女孩接口
-     * @return
-     */
-    GirlInterface getGirl();
-}
