--- src/main/java/com/study/patterns/factory/abstracts/po/BoyInterface.java
@@ -1,14 +1,0 @@
-package com.study.patterns.factory.abstracts.po;
-
-/**
- * 男孩接口
- *
- * @author Sky
- * @Date 2019-01-27 14:12.
- */
-public interface BoyInterface {
-    /**
-     * 画一个男孩
-     */
-    void drawMan();
-}
