--- src/main/java/com/study/patterns/factory/abstracts/po/GirlInterface.java
@@ -1,14 +1,0 @@
-package com.study.patterns.factory.abstracts.po;
-
-/**
- * 女孩接口
- *
- * @author Sky
- * @Date 2019-01-27 14:13.
- */
-public interface GirlInterface {
-    /**
-     * 画一个女孩
-     */
-    void drawWomen();
-}
