--- src/main/java/com/study/patterns/factory/methods/po/HairInterface.java
@@ -1,16 +1,0 @@
-package com.study.patterns.factory.methods.po;
-
-/**
- * 发型接口
- *
- * @author Sky
- * @Date 2019-01-27 13:00.
- */
-public interface HairInterface {
-
-    /**
-     * 画发型
-     */
-    void draw();
-
-}
