--- src/main/java/com/study/patterns/factory/methods/type.properties
@@ -1,3 +1,0 @@
-left = com.study.patterns.factory.methods.po.HairLeft
-right = com.study.patterns.factory.methods.po.HairRight
-in = com.study.patterns.factory.methods.po.HairIn
